Plugin Message sending works.

+++ src/GuiPluginMessages.cpp
@@ -7,7 +7,7 @@
 bool GuiPluginMessages::m_bDraw=true;
 
 
-int GuiPluginMessages::m_msg_sender_plugin_id=0;
-int GuiPluginMessages::m_msg_target_plugin_id=1;
-int GuiPluginMessages::m_msg_message=2;
-int GuiPluginMessages::m_msg_param=3; //refcons etc
+size_t GuiPluginMessages::m_msg_sender_plugin_id=0;
+size_t GuiPluginMessages::m_msg_target_plugin_id=1;
+size_t GuiPluginMessages::m_msg_message=2;
+size_t GuiPluginMessages::m_msg_param=3; //refcons etc
